+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Added the Wolf-hound role. ([#50](https://github.com/hjwylde/werewolf/issues/50))
 * Added a `version` command. ([#84](https://github.com/hjwylde/werewolf/issues/84))
 
+*Revisions*
+
+* Renamed the Villager role to Simple Villager.
+
 #### v0.4.4.1
 
 *Revisions*